BatchUpdate.cbl
Master File Batch Update 
The transaction file is provided as TRANSACTIONS.TXT. The records in the transaction file are in the same ascending sequence by key field (Part Number) as the master inventory file. The master inventory file is provided as INVENT4.txt.  The Suppliers file is provided as SUPPLIERS.txt. The resulting master file from this update must be named INVENT5.txt