Create Online update
Online Update Program 
The Inventory Master file, now an Indexed Sequential File will be updated using online (interactive) transactions. You must use a Screen Section in your program to enter the transaction data and apply it to the master file. 
The transactions for this update are provided below in Online Transactions

These are the transactions that you must use in the online update of the Inventory Master file once it has been converted to an Indexed Sequential file. The 6 transactions must be handled in the same order as given below.
S represents the sale of the number of units and R represents the receipt of a number of units
31141  S  10
31131  S  800
21112  S  150
21117  S  100
11141  R  100
21141  R  50
The program should prompt the user whether a transaction is to be entered and, if so, prompt for the inventory number, the type of transaction (S or R) and the amount.
A stated in an earlier requirement note, you must use the Screen Section to handle all view layer processes. I leave the specific design of that structure to you. However, you must capture screen shots for all six transactions noted above and include them in your team submission.