Create Inventory report
Inventory Report Program (from project 3)
You must run the report program from Project 3. However you will have to alter the reference in the Environment Division to ensure you are referring to the correct version of the file.
In this program, you must use the COPY command for the record structure of the Inventory Record.
You must use the CALL command to execute the module to compute the value of the inventory item.
Re-Order Report
You must compose the structure of the re-order report, but it must have the Inventory Number, Inventory Name, Quantity on hand and the Supplier Name. The Supplier Name must be retrieved from the Suppliers Index Sequential File (SUPPLIERI) created as an indexed sequential file (see above -- Master File Conversion / Suppliers file conversion)